+++ extensions/video/src/bindbc/libvlc/package.d
@@ -1,0 +1,9 @@
+module bindbc.libvlc;
+
+public
+{
+    import bindbc.libvlc.types;
+    import bindbc.libvlc.funcs;
+    import bindbc.libvlc.nativefuncs;
+    import bindbc.libvlc.loader;
+}